adding the [hopefully] final draft
